refactor: P2 TK companion 抽取（功能不变）
把 Chat Completions / Messages / failover / auth / gemini v1beta / pricing
lookup 等上游形文件中的 TK 块外提到 *_tk_*.go，主文件只留薄调用点，降低
upstream merge 冲突面。同步 protocol-routing SSOT 与 gateway-tk sentinel。
no-web-impact
sentinel-registry-reviewed
