+++ backend/internal/handler/auth_handler.go
@@ -1,7 +1,6 @@
 package handler
 
 import (
-	"context"
 	"log/slog"
 	"strings"
 	"sync"
@@ -149,34 +148,6 @@ func respondWithTokenPair(c *gin.Context, authService *service.AuthService, user
 		TokenType:    "Bearer",
 		User:         dto.UserFromService(user),
 	})
-}
-
-func (h *AuthHandler) ensureBackendModeAllowsUser(ctx context.Context, user *service.User) error {
-	if user == nil {
-		return infraerrors.Unauthorized("INVALID_USER", "user not found")
-	}
-	if h == nil || !h.isBackendModeEnabled(ctx) || user.IsAdmin() {
-		return nil
-	}
-	return infraerrors.Forbidden("BACKEND_MODE_ADMIN_ONLY", "Backend mode is active. Only admin login is allowed.")
-}
-
-func (h *AuthHandler) ensureBackendModeAllowsNewUserLogin(ctx context.Context) error {
-	if h == nil || !h.isBackendModeEnabled(ctx) {
-		return nil
-	}
-	return infraerrors.Forbidden("BACKEND_MODE_ADMIN_ONLY", "Backend mode is active. Only admin login is allowed.")
-}
-
-func (h *AuthHandler) isBackendModeEnabled(ctx context.Context) bool {
-	if h == nil || h.settingSvc == nil {
-		return false
-	}
-	settings, err := h.settingSvc.GetPublicSettings(ctx)
-	if err == nil && settings != nil {
-		return settings.BackendModeEnabled
-	}
-	return h.settingSvc.IsBackendModeEnabled(ctx)
 }
 
 // Register handles user registration
