+++ backend/internal/handler/failover_loop.go
@@ -2,13 +2,10 @@ package handler
 
 import (
 	"context"
-	"encoding/json"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/tidwall/gjson"
 
 	"github.com/Wei-Shaw/sub2api/internal/pkg/logger"
 	"github.com/Wei-Shaw/sub2api/internal/service"
@@ -45,17 +42,7 @@ const (
 	// Service 层在 SingleAccountRetry 模式下已做充分原地重试（最多 3 次、总等待 30s），
 	// Handler 层只需短暂间隔后重新进入 Service 层即可。
 	singleAccountBackoffDelay = 2 * time.Second
-	// maxProfitVetoAttempts 单次请求内允许的分组利润门终检否决次数上限。
-	// 利润否决不产生上游请求，因此不会推进 SwitchCount；没有独立上限的话，
-	// 「选号 → 终检否决 → 重选」在候选池与账号快照短暂不一致时可以空转很久。
-	// 取值与 maxAccountSwitches 默认值一致：混合定价的大分组仍有充分重选机会，
-	// 同时把整池越线时的无谓选号开销限制在常数级。
-	maxProfitVetoAttempts = 10
 )
-
-// profitVetoExhaustedMessage 是利润否决次数耗尽时返回给客户端的文案。
-// 语义上等同于「无可用账号」：候选账号都不满足分组的利润约束。
-const profitVetoExhaustedMessage = "No available accounts: all candidates rejected by group profit control"
 
 func sameAccountRetryDelayFor(failoverErr *service.UpstreamFailoverError, retryCount int) time.Duration {
 	if failoverErr == nil {
@@ -170,42 +157,6 @@ func (s *FailoverState) sleep(ctx context.Context, delay time.Duration) bool {
 	return s.sleeper(ctx, delay)
 }
 
-// RecordProfitVeto 记录一次分组利润门终检否决：把账号加入排除列表（同时登记到
-// 利润否决集，使其不被 503 退避分支清掉）并递增否决计数。
-//
-// 返回 FailoverContinue 表示调用方可以继续重选下一个账号；返回 FailoverExhausted
-// 表示本次请求的利润否决次数已达上限，调用方应按「无可用账号」终止，
-// 不得继续 continue。
-func (s *FailoverState) RecordProfitVeto(accountID int64) FailoverAction {
-	s.FailedAccountIDs[accountID] = struct{}{}
-	if s.profitVetoedAccountIDs == nil {
-		s.profitVetoedAccountIDs = make(map[int64]struct{})
-	}
-	s.profitVetoedAccountIDs[accountID] = struct{}{}
-	s.profitVetoCount++
-	if s.profitVetoCount >= maxProfitVetoAttempts {
-		return FailoverExhausted
-	}
-	return FailoverContinue
-}
-
-// ProfitVetoCount 返回本次请求累计的利润否决次数（供日志使用）。
-func (s *FailoverState) ProfitVetoCount() int { return s.profitVetoCount }
-
-// allExclusionsAreProfitVetoed 判断排除列表是否已全部由利润门否决贡献。
-// 此时清空 FailedAccountIDs 会被原样恢复，退避重试不会带来任何新候选。
-func (s *FailoverState) allExclusionsAreProfitVetoed() bool {
-	if len(s.profitVetoedAccountIDs) == 0 || len(s.FailedAccountIDs) == 0 {
-		return false
-	}
-	for id := range s.FailedAccountIDs {
-		if _, ok := s.profitVetoedAccountIDs[id]; !ok {
-			return false
-		}
-	}
-	return true
-}
-
 // HandleFailoverError 处理 UpstreamFailoverError，返回下一步动作。
 // 包含：缓存计费判断、同账号重试、临时封禁、切换计数、Antigravity 延时。
 func (s *FailoverState) HandleFailoverError(
@@ -294,12 +245,8 @@ func (s *FailoverState) HandleFailoverError(
 		zap.Int("max_switches", s.MaxSwitches),
 	)
 
-	// Antigravity 平台换号线性递增延时
-	if platform == service.PlatformAntigravity {
-		delay := time.Duration(s.SwitchCount-1) * time.Second
-		if !s.sleep(ctx, delay) {
-			return FailoverCanceled
-		}
+	if action := s.applyAntigravitySwitchDelay(ctx, platform); action != FailoverContinue {
+		return action
 	}
 
 	return FailoverContinue
@@ -325,14 +272,7 @@ func (s *FailoverState) HandleSelectionExhausted(ctx context.Context, thinPoolEx
 		(s.LastFailoverErr != nil && s.LastFailoverErr.StatusCode == http.StatusServiceUnavailable))
 	if retryable && s.SwitchCount <= s.MaxSwitches {
 
-		// 排除列表全由利润门否决贡献时，清空后会被原样恢复：退避重试拿不到
-		// 任何新候选，而利润否决不推进 SwitchCount，退避条件将永远成立。
-		// 这里直接判定耗尽，避免每 2s 空转一轮的活锁。
-		if s.allExclusionsAreProfitVetoed() {
-			logger.FromContext(ctx).Warn("gateway.failover_selection_exhausted_by_profit_veto",
-				zap.Int("profit_veto_count", s.profitVetoCount),
-				zap.Int("excluded_accounts", len(s.FailedAccountIDs)),
-			)
+		if s.maybeExhaustByProfitVeto(ctx) {
 			return FailoverExhausted
 		}
 
@@ -351,11 +291,7 @@ func (s *FailoverState) HandleSelectionExhausted(ctx context.Context, thinPoolEx
 			zap.Int("max_switches", s.MaxSwitches),
 		)
 		s.FailedAccountIDs = make(map[int64]struct{})
-		// 利润门否决的账号不参与退避重试的解除：判定依据（冻结的下游倍率）在
-		// 同一请求内不变，放它们回池只会被再次否决。
-		for id := range s.profitVetoedAccountIDs {
-			s.FailedAccountIDs[id] = struct{}{}
-		}
+		s.restoreProfitVetoExclusions()
 		return FailoverContinue
 	}
 	return FailoverExhausted
@@ -391,20 +327,4 @@ func sleepWithContext(ctx context.Context, d time.Duration) bool {
 	case <-time.After(d):
 		return true
 	}
-}
-
-// TokenKey middleware shape: {"code":"INSUFFICIENT_BALANCE","message":"..."}.
-func looksLikeStructuredErrorJSON(body []byte) bool {
-	if len(body) == 0 {
-		return false
-	}
-	if !json.Valid(body) {
-		return false
-	}
-	if gjson.GetBytes(body, "error").IsObject() {
-		return true
-	}
-	code := strings.TrimSpace(gjson.GetBytes(body, "code").String())
-	msg := strings.TrimSpace(gjson.GetBytes(body, "message").String())
-	return code != "" && msg != ""
 }
