+++ backend/internal/handler/gateway_handler.go
@@ -2,8 +2,6 @@ package handler
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"math"
@@ -18,7 +16,6 @@ import (
 	"github.com/Wei-Shaw/sub2api/internal/engine/protocolrouter"
 	"github.com/Wei-Shaw/sub2api/internal/pkg/antigravity"
 	"github.com/Wei-Shaw/sub2api/internal/pkg/claude"
-	"github.com/Wei-Shaw/sub2api/internal/pkg/ctxkey"
 	pkgerrors "github.com/Wei-Shaw/sub2api/internal/pkg/errors"
 	"github.com/Wei-Shaw/sub2api/internal/pkg/geminicli"
 	"github.com/Wei-Shaw/sub2api/internal/pkg/ip"
@@ -334,305 +331,8 @@ func (h *GatewayHandler) Messages(c *gin.Context) {
 	}
 
 	if platform == service.PlatformGemini {
-		fs := NewFailoverState(h.maxAccountSwitchesGemini, hasBoundSession)
-
-		// 单账号分组提前设置 SingleAccountRetry 标记，让 Service 层首次 503 就不设模型限流标记。
-		// 避免单账号分组收到 503 (MODEL_CAPACITY_EXHAUSTED) 时设 29s 限流，导致后续请求连续快速失败。
-		if h.gatewayService.IsSingleAntigravityAccountGroup(c.Request.Context(), apiKey.GroupID) {
-			ctx := service.WithSingleAccountRetry(c.Request.Context(), true, h.metadataBridgeEnabled())
-			c.Request = c.Request.WithContext(ctx)
-		}
-
-		for {
-			routingStart := time.Now()
-			selection, err := h.gatewayService.SelectAccountWithLoadAwareness(c.Request.Context(), apiKey.GroupID, sessionKey, reqModel, fs.FailedAccountIDs, "", int64(0)) // Gemini 不使用会话限制
-			if err != nil {
-				if len(fs.FailedAccountIDs) == 0 {
-					if h.tkWriteDeprecatedAnthropicModelIfApplicable(c, err, reqModel, reqLog) {
-						return
-					}
-					if h.tkWriteUnsupportedModelIfApplicable(c, err, reqModel, streamStarted, reqLog) {
-						return
-					}
-					markOpsRoutingCapacityLimitedIfNoAvailable(c, err)
-					reqLog.Warn("gateway.select_account_no_available",
-						zap.String("model", reqModel),
-						zap.Int64p("group_id", apiKey.GroupID),
-						zap.String("platform", platform),
-						zap.Error(err),
-					)
-					h.handleStreamingAwareError(c, tkNoAvailableAccounts(c), "api_error", "No available accounts: "+err.Error(), streamStarted)
-					return
-				}
-				action := fs.HandleSelectionExhausted(c.Request.Context(), errors.Is(err, service.ErrThinPoolAllExcluded))
-				switch action {
-				case FailoverContinue:
-					ctx := service.WithSingleAccountRetry(c.Request.Context(), true, h.metadataBridgeEnabled())
-					c.Request = c.Request.WithContext(ctx)
-					continue
-				case FailoverCanceled:
-					failoverClientGone(c)
-					return
-				default: // FailoverExhausted
-					if fs.LastFailoverErr != nil {
-						h.handleFailoverExhausted(c, fs.LastFailoverErr, service.PlatformGemini, streamStarted)
-					} else {
-						h.handleFailoverExhaustedSimple(c, 502, streamStarted)
-					}
-					return
-				}
-			}
-			account := selection.Account
-			setOpsSelectedAccountFrom(c, account)
-
-			// 检查请求拦截（预热请求、SUGGESTION MODE等）
-			if account.IsInterceptWarmupEnabled() {
-				interceptType := detectInterceptType(body, reqModel, parsedReq.MaxTokens, isClaudeCodeClient)
-				if interceptType != InterceptTypeNone {
-					if selection.Acquired && selection.ReleaseFunc != nil {
-						selection.ReleaseFunc()
-					}
-					if reqStream {
-						sendMockInterceptStream(c, reqModel, interceptType, account)
-					} else {
-						sendMockInterceptResponse(c, reqModel, interceptType, account)
-					}
-					return
-				}
-			}
-
-			// 3. 获取账号并发槽位
-			accountReleaseFunc := selection.ReleaseFunc
-			if !selection.Acquired {
-				if selection.WaitPlan == nil {
-					markOpsRoutingCapacityLimited(c)
-					reqLog.Warn("gateway.select_account_no_slot_no_wait_plan",
-						zap.Int64("account_id", account.ID),
-						zap.String("model", reqModel),
-						zap.String("platform", platform),
-					)
-					h.handleStreamingAwareError(c, tkNoAvailableAccounts(c), "api_error", "No available accounts", streamStarted)
-					return
-				}
-				accountWaitCounted := false
-				canWait, err := h.concurrencyHelper.IncrementAccountWaitCount(c.Request.Context(), account.ID, selection.WaitPlan.MaxWaiting)
-				if err != nil {
-					reqLog.Warn("gateway.account_wait_counter_increment_failed", zap.Int64("account_id", account.ID), zap.Error(err))
-				} else if !canWait {
-					reqLog.Info("gateway.account_wait_queue_full",
-						zap.Int64("account_id", account.ID),
-						zap.Int("max_waiting", selection.WaitPlan.MaxWaiting),
-					)
-					h.handleStreamingAwareError(c, http.StatusTooManyRequests, "rate_limit_error", "Too many pending requests, please retry later", streamStarted)
-					return
-				}
-				if err == nil && canWait {
-					accountWaitCounted = true
-				}
-				releaseWait := func() {
-					if accountWaitCounted {
-						h.concurrencyHelper.DecrementAccountWaitCount(c.Request.Context(), account.ID)
-						accountWaitCounted = false
-					}
-				}
-
-				accountReleaseFunc, err = h.concurrencyHelper.AcquireAccountSlotWithWaitTimeout(
-					c,
-					account.ID,
-					selection.WaitPlan.MaxConcurrency,
-					selection.WaitPlan.Timeout,
-					reqStream,
-					&streamStarted,
-				)
-				if err != nil {
-					reqLog.Warn("gateway.account_slot_acquire_failed", zap.Int64("account_id", account.ID), zap.Error(err))
-					releaseWait()
-					h.handleConcurrencyError(c, err, "account", streamStarted)
-					return
-				}
-				// Slot acquired: no longer waiting in queue.
-				releaseWait()
-			}
-			// 终检与准入后绑定使用选号结果携带的门（见 responses 同名注释）。
-			admissionCtx := service.ContextWithSelectionProfitGate(c.Request.Context(), selection)
-			latest, vetoed, reason := h.gatewayService.GatewayProfitControlVetoLatest(admissionCtx, account)
-			if vetoed {
-				if accountReleaseFunc != nil {
-					accountReleaseFunc()
-				}
-				reqLog.Debug("gateway.account_slot_profit_vetoed", zap.Int64("account_id", account.ID), zap.String("reason", reason))
-				if fs.RecordProfitVeto(account.ID) == FailoverExhausted {
-					reqLog.Warn("gateway.profit_veto_attempts_exhausted", zap.Int("profit_veto_count", fs.ProfitVetoCount()))
-					markOpsRoutingCapacityLimited(c)
-					h.handleStreamingAwareError(c, http.StatusServiceUnavailable, "api_error", profitVetoExhaustedMessage, streamStarted)
-					return
-				}
-				continue
-			}
-			account = latest
-			selection.Account = latest
-			// 等待路径保持既有 eager 绑定（无门时 helper 直接绑定）；调度器已
-			// 抢槽的直达路径无门时由选号内部绑定，这里只在门下补准入后绑定。
-			if selection.ProfitGateActive() || !selection.Acquired {
-				if err := h.gatewayService.BindStickySessionAfterProfitAdmission(admissionCtx, apiKey.GroupID, sessionKey, account.ID); err != nil {
-					reqLog.Warn("gateway.bind_sticky_session_after_profit_admission_failed", zap.Int64("account_id", account.ID), zap.Error(err))
-				}
-			}
-			// 账号槽位/等待计数需要在超时或断开时安全回收
-			accountReleaseFunc = wrapReleaseOnDone(c.Request.Context(), accountReleaseFunc)
-
-			tkRecordRoutingLatency(c, routingStart)
-
-			// 转发请求 - 根据账号平台分流
-			var result *service.ForwardResult
-			requestCtx := c.Request.Context()
-			if fs.SwitchCount > 0 {
-				requestCtx = service.WithAccountSwitchCount(requestCtx, fs.SwitchCount, h.metadataBridgeEnabled())
-			}
-			// 记录 Forward 前已写入字节数，Forward 后若增加则说明 SSE 内容已发，禁止 failover
-			writerSizeBeforeForward := c.Writer.Size()
-			forwardStart := time.Now()
-			if account.Platform == service.PlatformAntigravity {
-				result, err = h.antigravityGatewayService.ForwardGemini(
-					requestCtx,
-					c,
-					account,
-					reqModel,
-					"generateContent",
-					reqStream,
-					body,
-					hasBoundSession,
-					service.WithForwardGeminiSession(derefGroupID(apiKey.GroupID), sessionKey),
-				)
-			} else {
-				// TK: 让 GeminiMessagesCompatService.Forward 能拿到 *Group 做
-				// 分组级 Claude→Gemini 映射 (gin_gemini_dispatch_tk.go)。
-				c.Set(service.TKGeminiDispatchGroupContextKey, apiKey.Group)
-				result, err = h.geminiCompatService.Forward(requestCtx, c, account, body)
-			}
-			tkRecordForwardResponseTail(c, forwardStart)
-			if accountReleaseFunc != nil {
-				accountReleaseFunc()
-			}
-			if err != nil {
-				var failoverErr *service.UpstreamFailoverError
-				if errors.As(err, &failoverErr) {
-					// 流式内容已写入客户端，无法撤销，禁止 failover 以防止流拼接腐化
-					if c.Writer.Size() != writerSizeBeforeForward {
-						h.handleFailoverExhausted(c, failoverErr, service.PlatformGemini, true)
-						return
-					}
-					action := fs.HandleFailoverError(c.Request.Context(), h.gatewayService, account.ID, account.Platform, account.GetPoolModeRetryCount(), failoverErr)
-					switch action {
-					case FailoverContinue:
-						continue
-					case FailoverExhausted:
-						h.handleFailoverExhausted(c, fs.LastFailoverErr, service.PlatformGemini, streamStarted)
-						return
-					case FailoverCanceled:
-						failoverClientGone(c)
-						return
-					}
-				}
-				upstreamErrorAlreadyCommunicated := gatewayForwardErrorAlreadyCommunicated(c, writerSizeBeforeForward, err)
-				wroteFallback := false
-				if !upstreamErrorAlreadyCommunicated {
-					wroteFallback = h.ensureForwardErrorResponseForError(c, err, streamStarted)
-				}
-				forwardFailedFields := []zap.Field{
-					zap.Int64("account_id", account.ID),
-					zap.String("account_name", account.Name),
-					zap.String("account_platform", account.Platform),
-					zap.Bool("fallback_error_response_written", wroteFallback),
-					zap.Bool("upstream_error_response_already_written", upstreamErrorAlreadyCommunicated),
-					zap.Error(err),
-				}
-				if account.Proxy != nil {
-					forwardFailedFields = append(forwardFailedFields,
-						zap.Int64("proxy_id", account.Proxy.ID),
-						zap.String("proxy_name", account.Proxy.Name),
-						zap.String("proxy_host", account.Proxy.Host),
-						zap.Int("proxy_port", account.Proxy.Port),
-					)
-				} else if account.ProxyID != nil {
-					forwardFailedFields = append(forwardFailedFields, zap.Int64p("proxy_id", account.ProxyID))
-				}
-				reqLog.Error("gateway.forward_failed", forwardFailedFields...)
-				// TK: passive availability failure tap (R-004 — extracts upstream HTTP status from UpstreamFailoverError)
-				TkRecordFailureFromErr(h.gatewayService, c.Request.Context(), account.Platform, reqModel, account.ID, err)
-				return
-			}
-
-			// RPM 计数递增（Forward 成功后）
-			// 注意：TOCTOU 竞态是已知且可接受的设计权衡，与其它调度 soft-limit 模式一致。
-			// 在高并发下可能短暂超出 RPM 限制，但不会导致请求失败。
-			if (account.IsAnthropicOAuthOrSetupToken() || account.IsKiro()) && account.GetBaseRPM() > 0 {
-				if err := h.gatewayService.IncrementAccountRPM(c.Request.Context(), account.ID); err != nil {
-					reqLog.Warn("gateway.rpm_increment_failed", zap.Int64("account_id", account.ID), zap.Error(err))
-				}
-			}
-			setOpsForwardResultContext(c, result.UpstreamModel, reqModel)
-			setOpsClaudeUsageContext(c, result.Usage)
-
-			// 捕获请求信息（用于异步记录，避免在 goroutine 中访问 gin.Context）
-			userAgent := c.GetHeader("User-Agent")
-			clientIP := ip.GetClientIP(c)
-			requestPayloadHash := service.HashUsageRequestPayload(body)
-			inboundEndpoint := GetInboundEndpoint(c)
-			upstreamEndpoint := GetUpstreamEndpoint(c, account.Platform)
-
-			if result.ReasoningEffort == nil {
-				result.ReasoningEffort = service.NormalizeClaudeOutputEffort(parsedReq.OutputEffort)
-			}
-			// 国产模型 thinking-enabled 默认 effort 填充：Kimi/GLM/MiniMax 这些不支持 effort 档位的
-			// passback-required 上游，仅要 thinking 启用且 OutputEffort 未明确传递时，在 usage_log 写 "high"
-			// 避免该字段长期为 NULL（详见 DefaultEffortForThinkingEnabled 文档）。
-			if result.ReasoningEffort == nil && parsedReq.ThinkingEnabled {
-				protocolModel := result.UpstreamModel
-				if protocolModel == "" {
-					protocolModel = result.Model
-				}
-				result.ReasoningEffort = service.DefaultEffortForThinkingEnabled(protocolModel)
-			}
-
-			// 使用量记录通过有界 worker 池提交，避免请求热路径创建无界 goroutine。
-			// ForceCacheBilling 提前拍成标量，避免 worker 闭包保活 failover 状态里的响应体。
-			forceCacheBilling := fs.ForceCacheBilling
-			quotaPlatform := service.QuotaPlatform(c.Request.Context(), apiKey)
-			sessionID := service.ExtractClientSessionID(c)
-			gatewayLatencyMs := tkSnapshotGatewayTransferLatencyMs(c)
-			h.submitUsageRecordTask(c.Request.Context(), func(ctx context.Context) {
-				if err := h.gatewayService.RecordUsage(ctx, &service.RecordUsageInput{
-					Result:             result,
-					QuotaPlatform:      quotaPlatform,
-					APIKey:             apiKey,
-					User:               apiKey.User,
-					Account:            account,
-					Subscription:       subscription,
-					PricingAt:          pricingAt,
-					InboundEndpoint:    inboundEndpoint,
-					UpstreamEndpoint:   upstreamEndpoint,
-					UserAgent:          userAgent,
-					IPAddress:          clientIP,
-					SessionID:          sessionID,
-					RequestPayloadHash: requestPayloadHash,
-					ForceCacheBilling:  forceCacheBilling,
-					APIKeyService:      h.apiKeyService,
-					GatewayLatencyMs:   gatewayLatencyMs,
-					ChannelUsageFields: clientRequestedUsageFields(c, channelMapping, reqModel, result.UpstreamModel),
-				}); err != nil {
-					logger.L().With(
-						zap.String("component", "handler.gateway.messages"),
-						zap.Int64("user_id", subject.UserID),
-						zap.Int64("api_key_id", apiKey.ID),
-						zap.Any("group_id", apiKey.GroupID),
-						zap.String("model", reqModel),
-						zap.Int64("account_id", account.ID),
-					).Error("gateway.record_usage_failed", zap.Error(err))
-				}
-			})
-			return
-		}
+		h.tkMessagesGeminiPlatform(c, reqLog, apiKey, subject, subscription, sessionKey, hasBoundSession, reqModel, reqStream, body, parsedReq, isClaudeCodeClient, streamStarted, platform, pricingAt, channelMapping)
+		return
 	}
 
 	currentAPIKey := apiKey
@@ -956,50 +656,14 @@ func (h *GatewayHandler) Messages(c *gin.Context) {
 					return
 				}
 
-				var promptTooLongErr *service.PromptTooLongError
-				if errors.As(err, &promptTooLongErr) {
-					reqLog.Warn("gateway.prompt_too_long_from_antigravity",
-						zap.Any("current_group_id", currentAPIKey.GroupID),
-						zap.Any("fallback_group_id", fallbackGroupID),
-						zap.Bool("fallback_used", fallbackUsed),
-					)
-					if !fallbackUsed && fallbackGroupID != nil && *fallbackGroupID > 0 {
-						fallbackGroup, err := h.gatewayService.ResolveGroupByID(c.Request.Context(), *fallbackGroupID)
-						if err != nil {
-							reqLog.Warn("gateway.resolve_fallback_group_failed", zap.Int64("fallback_group_id", *fallbackGroupID), zap.Error(err))
-							_ = h.antigravityGatewayService.WriteMappedClaudeError(c, account, promptTooLongErr.StatusCode, promptTooLongErr.RequestID, promptTooLongErr.Body)
-							return
-						}
-						if fallbackGroup.Platform != service.PlatformAnthropic ||
-							fallbackGroup.SubscriptionType == service.SubscriptionTypeSubscription ||
-							fallbackGroup.FallbackGroupIDOnInvalidRequest != nil {
-							reqLog.Warn("gateway.fallback_group_invalid",
-								zap.Int64("fallback_group_id", fallbackGroup.ID),
-								zap.String("fallback_platform", fallbackGroup.Platform),
-								zap.String("fallback_subscription_type", fallbackGroup.SubscriptionType),
-							)
-							_ = h.antigravityGatewayService.WriteMappedClaudeError(c, account, promptTooLongErr.StatusCode, promptTooLongErr.RequestID, promptTooLongErr.Body)
-							return
-						}
-						fallbackAPIKey := cloneAPIKeyWithGroup(apiKey, fallbackGroup)
-						if err := h.billingCacheService.CheckBillingEligibility(c.Request.Context(), fallbackAPIKey.User, fallbackAPIKey, fallbackGroup, nil, service.PlatformFromAPIKey(fallbackAPIKey)); err != nil {
-							status, code, message, retryAfter := billingErrorDetails(err)
-							if retryAfter > 0 {
-								c.Header("Retry-After", strconv.Itoa(retryAfter))
-							}
-							h.handleStreamingAwareError(c, status, code, message, streamStarted)
-							return
-						}
-						// 兜底重试按"直接请求兜底分组"处理：清除强制平台，允许按分组平台调度
-						ctx := context.WithValue(c.Request.Context(), ctxkey.ForcePlatform, "")
-						c.Request = c.Request.WithContext(ctx)
-						currentAPIKey = fallbackAPIKey
+				if newAPIKey, retry, handled := h.tkHandleMessagesPromptTooLongFallback(c, reqLog, err, account, apiKey, currentAPIKey, fallbackGroupID, fallbackUsed, streamStarted); handled {
+					if retry {
+						currentAPIKey = newAPIKey
 						currentSubscription = nil
 						fallbackUsed = true
 						retryWithFallback = true
 						break
 					}
-					_ = h.antigravityGatewayService.WriteMappedClaudeError(c, account, promptTooLongErr.StatusCode, promptTooLongErr.RequestID, promptTooLongErr.Body)
 					return
 				}
 				var failoverErr *service.UpstreamFailoverError
@@ -1978,224 +1642,6 @@ func (h *GatewayHandler) CountTokens(c *gin.Context) {
 	// 选号 + 转发，带账号 failover（与 /v1/messages 一致；pool_mode stub 自动走
 	// 池内轮换）。见 gateway_handler_tk_count_tokens_failover.go。
 	h.forwardCountTokensWithFailover(c, reqLog, apiKey, sessionHash, parsedReq)
-}
-
-// InterceptType 表示请求拦截类型
-type InterceptType int
-
-const (
-	InterceptTypeNone              InterceptType = iota
-	InterceptTypeWarmup                          // 预热请求（返回 "New Conversation"）
-	InterceptTypeSuggestionMode                  // SUGGESTION MODE（返回空字符串）
-	InterceptTypeMaxTokensOneHaiku               // max_tokens=1 + haiku 探测请求（返回 "#"）
-)
-
-// isHaikuModel 检查模型名称是否包含 "haiku"（大小写不敏感）
-func isHaikuModel(model string) bool {
-	return strings.Contains(strings.ToLower(model), "haiku")
-}
-
-// isMaxTokensOneHaikuRequest 检查是否为 max_tokens=1 + haiku 模型的探测请求
-// 这类请求用于 Claude Code 验证 API 连通性（流式/非流式均会出现，如 cc-switch v3.9.0 起的健康检查探测为流式）
-// 条件：max_tokens == 1 且 model 包含 "haiku"
-func isMaxTokensOneHaikuRequest(model string, maxTokens int) bool {
-	return maxTokens == 1 && isHaikuModel(model)
-}
-
-// detectInterceptType 检测请求是否需要拦截，返回拦截类型
-// 参数说明：
-//   - body: 请求体字节
-//   - model: 请求的模型名称
-//   - maxTokens: max_tokens 值
-//   - isClaudeCodeClient: 是否已通过 Claude Code 客户端校验
-func detectInterceptType(body []byte, model string, maxTokens int, isClaudeCodeClient bool) InterceptType {
-	// 优先检查 max_tokens=1 + haiku 探测请求（流式/非流式均适用）
-	if isClaudeCodeClient && isMaxTokensOneHaikuRequest(model, maxTokens) {
-		return InterceptTypeMaxTokensOneHaiku
-	}
-
-	// 快速检查：如果不包含任何关键字，直接返回
-	bodyStr := string(body)
-	hasSuggestionMode := strings.Contains(bodyStr, "[SUGGESTION MODE:")
-	hasWarmupKeyword := strings.Contains(bodyStr, "title") || strings.Contains(bodyStr, "Warmup")
-
-	if !hasSuggestionMode && !hasWarmupKeyword {
-		return InterceptTypeNone
-	}
-
-	// 解析请求（只解析一次）
-	var req struct {
-		Messages []struct {
-			Role    string `json:"role"`
-			Content []struct {
-				Type string `json:"type"`
-				Text string `json:"text"`
-			} `json:"content"`
-		} `json:"messages"`
-		System []struct {
-			Text string `json:"text"`
-		} `json:"system"`
-	}
-	if err := json.Unmarshal(body, &req); err != nil {
-		return InterceptTypeNone
-	}
-
-	// 检查 SUGGESTION MODE（最后一条 user 消息）
-	if hasSuggestionMode && len(req.Messages) > 0 {
-		lastMsg := req.Messages[len(req.Messages)-1]
-		if lastMsg.Role == "user" && len(lastMsg.Content) > 0 &&
-			lastMsg.Content[0].Type == "text" &&
-			strings.HasPrefix(lastMsg.Content[0].Text, "[SUGGESTION MODE:") {
-			return InterceptTypeSuggestionMode
-		}
-	}
-
-	// 检查 Warmup 请求
-	if hasWarmupKeyword {
-		// 检查 messages 中的标题提示模式
-		for _, msg := range req.Messages {
-			for _, content := range msg.Content {
-				if content.Type == "text" {
-					if strings.Contains(content.Text, "Please write a 5-10 word title for the following conversation:") ||
-						content.Text == "Warmup" {
-						return InterceptTypeWarmup
-					}
-				}
-			}
-		}
-		// 检查 system 中的标题提取模式
-		for _, sys := range req.System {
-			if strings.Contains(sys.Text, "nalyze if this message indicates a new conversation topic. If it does, extract a 2-3 word title") {
-				return InterceptTypeWarmup
-			}
-		}
-	}
-
-	return InterceptTypeNone
-}
-
-// sendMockInterceptStream 发送流式 mock 响应（用于请求拦截）
-func sendMockInterceptStream(c *gin.Context, model string, interceptType InterceptType, account *service.Account) {
-	if account != nil {
-		setOpsSelectedAccountFrom(c, account)
-	}
-	c.Header("Content-Type", "text/event-stream")
-	c.Header("Cache-Control", "no-cache")
-	c.Header("Connection", "keep-alive")
-	c.Header("X-Accel-Buffering", "no")
-
-	// 根据拦截类型决定响应内容
-	var msgID string
-	var outputTokens int
-	var textDeltas []string
-
-	switch interceptType {
-	case InterceptTypeSuggestionMode:
-		msgID = generateRealisticMsgID()
-		outputTokens = 1
-		textDeltas = []string{""} // 空内容
-	default: // InterceptTypeWarmup
-		msgID = generateRealisticMsgID()
-		outputTokens = 2
-		textDeltas = []string{"New", " Conversation"}
-	}
-
-	// Build message_start event — field order matches real Anthropic API response.
-	messageStartJSON := `{"type":"message_start","message":{"model":` + strconv.Quote(model) + `,"id":` + strconv.Quote(msgID) + `,"type":"message","role":"assistant","content":[],"stop_reason":null,"stop_sequence":null,"stop_details":null,"usage":{"input_tokens":10,"cache_creation_input_tokens":0,"cache_read_input_tokens":0,"output_tokens":0}}}`
-
-	// Build events
-	events := []string{
-		`event: message_start` + "\n" + `data: ` + string(messageStartJSON),
-		`event: content_block_start` + "\n" + `data: {"type":"content_block_start","index":0,"content_block":{"type":"text","text":""}}`,
-	}
-
-	// Add text deltas
-	for _, text := range textDeltas {
-		deltaJSON := `{"type":"content_block_delta","index":0,"delta":{"type":"text_delta","text":` + strconv.Quote(text) + `}}`
-		events = append(events, `event: content_block_delta`+"\n"+`data: `+string(deltaJSON))
-	}
-
-	// Add final events
-	messageDeltaJSON := `{"type":"message_delta","delta":{"stop_reason":"end_turn","stop_sequence":null,"stop_details":null},"usage":{"output_tokens":` + strconv.Itoa(outputTokens) + `}}`
-
-	events = append(events,
-		`event: content_block_stop`+"\n"+`data: {"index":0,"type":"content_block_stop"}`,
-		`event: message_delta`+"\n"+`data: `+string(messageDeltaJSON),
-		`event: message_stop`+"\n"+`data: {"type":"message_stop"}`,
-	)
-
-	for _, event := range events {
-		_, _ = c.Writer.WriteString(event + "\n\n")
-		c.Writer.Flush()
-		time.Sleep(20 * time.Millisecond)
-	}
-}
-
-// generateRealisticMsgID 生成仿真的消息 ID（msg_01XXXXXXX 格式）
-// 格式与 Anthropic API 官方响应一致：msg_01 + 22 位 Base62 随机字符
-func generateRealisticMsgID() string {
-	const charset = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
-	const idLen = 22
-	randomBytes := make([]byte, idLen)
-	if _, err := rand.Read(randomBytes); err != nil {
-		return fmt.Sprintf("msg_01%d", time.Now().UnixNano())
-	}
-	b := make([]byte, idLen)
-	for i := range b {
-		b[i] = charset[int(randomBytes[i])%len(charset)]
-	}
-	return "msg_01" + string(b)
-}
-
-// sendMockInterceptResponse 发送非流式 mock 响应（用于请求拦截）
-func sendMockInterceptResponse(c *gin.Context, model string, interceptType InterceptType, account *service.Account) {
-	if account != nil {
-		setOpsSelectedAccountFrom(c, account)
-	}
-	var msgID, text, stopReason string
-	var outputTokens int
-
-	switch interceptType {
-	case InterceptTypeSuggestionMode:
-		msgID = generateRealisticMsgID()
-		text = ""
-		outputTokens = 1
-		stopReason = "end_turn"
-	case InterceptTypeMaxTokensOneHaiku:
-		msgID = generateRealisticMsgID()
-		text = "#"
-		outputTokens = 1
-		stopReason = "max_tokens" // max_tokens=1 探测请求的 stop_reason 应为 max_tokens
-	default: // InterceptTypeWarmup
-		msgID = generateRealisticMsgID()
-		text = "New Conversation"
-		outputTokens = 2
-		stopReason = "end_turn"
-	}
-
-	// 构建完整的响应格式（与 Anthropic API 官方响应格式一致）
-	response := gin.H{
-		"model":         model,
-		"id":            msgID,
-		"type":          "message",
-		"role":          "assistant",
-		"content":       []gin.H{{"type": "text", "text": text}},
-		"stop_reason":   stopReason,
-		"stop_sequence": nil,
-		"stop_details":  nil,
-		"usage": gin.H{
-			"input_tokens":                10,
-			"cache_creation_input_tokens": 0,
-			"cache_read_input_tokens":     0,
-			"cache_creation": gin.H{
-				"ephemeral_5m_input_tokens": 0,
-				"ephemeral_1h_input_tokens": 0,
-			},
-			"output_tokens": outputTokens,
-		},
-	}
-
-	c.JSON(http.StatusOK, response)
 }
 
 // extractQuotaResetSeconds 从 quota 错误的 metadata 中提取 window_resets_at 并计算
